Fix spacing typo in the README
As I believe in the greater good of the English language and the development of documentation as the foundation for a tremendous future, I decided to fix a small typo in your README.

# What does that mean ?

In English, there is no space before a colon. This is a common mistake for French speakers since it is the case in this weird language. However you might want to know that it is one of the unique characteristic of French typesetting. You can read more about this [here](https://en.wikipedia.org/wiki/Sentence_spacing).

# Why would you do this ?

Each typo fixed saves another child in the world. It is a direct consequence of quantum physics and the butterfly effect. There's a great movie that talks about it where the main character goes back in time and changes little things and lives the consequences. It's called the Butterfly Effect.  

# Isn't this a useless PR ?

You know, I don't believe we should bring philosophy in the discussion but I don't think we have a choice. Whether or not this PR is useful or not is highly subjective and nobody should be judged so harshly, so fast.  
What if, for example, the President Of The United States decided to use Kamaï to rediscover Computer Science.  
What if he decided that it would become the universal programming language. Right there, overnight.  
By morning, everyone on earth would read this README and notice this spacing error. They would think "Wow, that's not a native speaker!" and since they're American, nukes would be involved.  

# Conclusion

Think of the children. Think of the mothers. Think of your ancestors. All thoses people that brought you there to live what you want to live.  
Do you want to deceive all of humanity with a spacing error ? I don't think you do. Thus I think you want to merge this PR.
Thank you.